change missing types to int|float

+++ src/Drivers/Color.php
@@ -62,6 +62,7 @@ abstract class Color
         };
     }
 
+    /** @return array<int> */
     protected function rgbaFromString(string $colorValue): array
     {
         // parse color string in hexidecimal format like #cccccc or cccccc or ccc
@@ -75,9 +76,9 @@ abstract class Color
 
         if (preg_match($hexPattern, $colorValue, $matches)) {
             $result = [];
-            $result[0] = strlen($matches[1]) == '1' ? hexdec($matches[1].$matches[1]) : hexdec($matches[1]);
-            $result[1] = strlen($matches[2]) == '1' ? hexdec($matches[2].$matches[2]) : hexdec($matches[2]);
-            $result[2] = strlen($matches[3]) == '1' ? hexdec($matches[3].$matches[3]) : hexdec($matches[3]);
+            $result[0] = strlen($matches[1]) == '1' ? (int) hexdec($matches[1].$matches[1]) : (int) hexdec($matches[1]);
+            $result[1] = strlen($matches[2]) == '1' ? (int) hexdec($matches[2].$matches[2]) : (int) hexdec($matches[2]);
+            $result[2] = strlen($matches[3]) == '1' ? (int) hexdec($matches[3].$matches[3]) : (int) hexdec($matches[3]);
             $result[3] = 1;
         } elseif (preg_match($rgbPattern, $colorValue, $matches)) {
             $result = [];
